renamed slots.py -> standard_slots.py
