+++ chatsky/slots/__init__.py
@@ -1,2 +1,2 @@
-from chatsky.slots.slots import RegexpSlot, RegexpGroupSlot, FunctionSlot
+from chatsky.slots.standard_slots import RegexpSlot, RegexpGroupSlot, FunctionSlot
 from chatsky.slots.base_slots import GroupSlot, ValueSlot
