+++ chatsky/slots/standard_slots.py
