+++ tests/slots/test_slot_types.py
@@ -3,7 +3,7 @@ from pydantic import ValidationError
 
 from chatsky.core import Message
 from chatsky.slots.base_slots import SlotNotExtracted, ExtractedValueSlot, GroupSlot, ExtractedGroupSlot
-from chatsky.slots.slots import RegexpSlot, FunctionSlot
+from chatsky.slots.standard_slots import RegexpSlot, FunctionSlot
 
 
 @pytest.mark.parametrize(
